fix: address PR feedback on activity card split test helpers
Rename zero-test helper modules out of Vitest discovery, colocate split
suites under current-activity-card/, and revert broadened lint allowlist debt.
